Rename component AB#12437
