+++ Apps/AdminWebClient/src/ClientApp/src/components/covidTreatmentAssessment/OptionDetails.vue
